Extract OSCALParties to a seperate file
